Fix SUM(amount) type mismatch and hardcode Stripe key in Dockerfile
PostgreSQL SUM() on BIGINT returns NUMERIC, not INT8. Added ::BIGINT
cast to both check_spend_limit implementations (handlers.rs and
billing.rs) to fix the "Database error" on credit topup.

Switched from ARG to ENV for NEXT_PUBLIC_* vars in the console
Dockerfile since Heroku build.config wasn't reliably forwarding them.
The Stripe publishable key is public by design (client-side JS).

Entire-Checkpoint: 9e7b4c12d32d

+++ bins/scrapix-api/src/auth/handlers.rs
@@ -1237,7 +1237,7 @@ async fn check_spend_limit(
     if let Some(limit) = limit {
         // Sum all top-ups this calendar month
         let spent: i64 = sqlx::query_scalar(
-            "SELECT COALESCE(SUM(amount), 0) FROM transactions \
+            "SELECT COALESCE(SUM(amount), 0)::BIGINT FROM transactions \
              WHERE account_id = $1 \
              AND type IN ('manual_topup', 'auto_topup') \
              AND created_at >= date_trunc('month', now())",
